Create NYC_Temp
This program analyzes the data of the rise in temperature over the last 200 years. First, it takes the February, August, and Annual average temperatures data and finds out the largest temperature and lowest temperature for each, in the entire period 1869-2017. Also, it gives the mean, median, mode and the standard deviation of the temperatures for each. Then it gives the slope,m, and y-intercept, b value for the line of best-fit equation, y = m(x)+b. After that, it plots the years and temperatures for February, August, and Annual average temperature and draws the line of best fit for all of them in a graph.